server/requirements.txt
mock
pyhamcrest
requests
nose
rednose
flask
